moved BookOpts to extras file
